TIP-841: Move the channel model to Akeneo\Channel
