+++ src/Pim/Bundle/CatalogBundle/Resources/config/entities.yml
@@ -22,8 +22,8 @@ parameters:
     pim_catalog.entity.attribute_requirement.class:        Pim\Bundle\CatalogBundle\Entity\AttributeRequirement
     pim_catalog.entity.family.class:                       Pim\Bundle\CatalogBundle\Entity\Family
     pim_catalog.entity.family_translation.class:           Pim\Bundle\CatalogBundle\Entity\FamilyTranslation
-    pim_catalog.entity.channel.class:                      Pim\Bundle\CatalogBundle\Entity\Channel
-    pim_catalog.entity.channel_translation.class:          Pim\Bundle\CatalogBundle\Entity\ChannelTranslation
+    pim_catalog.entity.channel.class:                      Akeneo\Channel\Component\Model\Channel
+    pim_catalog.entity.channel_translation.class:          Akeneo\Channel\Component\ModelTranslation
     pim_catalog.entity.currency.class:                     Pim\Bundle\CatalogBundle\Entity\Currency
     pim_catalog.entity.metric.class:                       Pim\Component\Catalog\Model\Metric
     pim_catalog.entity.product_price.class:                Pim\Component\Catalog\Model\ProductPrice
