+++ src/Pim/Component/User/Model/User.php
@@ -8,7 +8,7 @@ use DateTime;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 use Doctrine\ORM\Mapping as ORM;
-use Pim\Bundle\CatalogBundle\Entity\Channel;
+use Akeneo\Channel\Component\Model\Channel;
 use Pim\Component\Catalog\Model\ChannelInterface;
 use Pim\Component\User\Model\UserInterface;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
